+++ examples/cli-support/src/fxa_creds.rs
@@ -52,7 +52,7 @@ fn create_fxa_creds(path: &str, cfg: FxaConfig, scopes: &[&str]) -> Result<Firef
 }
 
 fn handle_oauth_flow(path: &str, acct: &FirefoxAccount, scopes: &[&str]) -> Result<()> {
-    let oauth_uri = acct.begin_oauth_flow(scopes, "fxa_creds")?;
+    let oauth_uri = acct.begin_oauth_flow(scopes, "fxa_creds", &["sync"])?;
 
     if open::that(&oauth_uri).is_err() {
         log::warn!("Failed to open a web browser D:");